+++ lib/services/forms/form_factory.dart
@@ -1,0 +1,10 @@
+
+import 'package:athlete_surveyor/models/interfaces/IGenericForm.dart';
+
+abstract class FormFactory {
+  /// After loading the form's [content], a 
+  IGenericForm createStudentForm();
+
+  IGenericForm createStaffForm();
+  
+}
